requirements.txt
# Linter - Static Code Quality - these are the current modules I am interested in using
flake8
flake8-annotations-complexity
flake8-breakpoint
flake8-builtins
flake8-comprehensions
flake8-eradicate
flake8-expression-complexity
pep8-naming
# flake8-isort
# flake8-logging-format
# flake8-requirements
# flake8-return # <- nice idea but buggy

# Formatters
black

# Code Modifications
vulture  # search for deadcode
isort    # Sort imports

# Add unit tests
pytest

# For Timezones on Windows. On linux, it will use the system binaries first
tzdata

# Add other packages here:
requests
pyyaml
click
appdirs
networkx
matplotlib


# Pandoc items
pandoc-fignos
pandoc-eqnos
pandoc-tablenos
pandoc-secnos
